use memcpy rather than hand-made loops. Use uint16_t as size type rather than uint8_t.

+++ src/common/buffer.h
@@ -36,12 +36,12 @@ public:
 	void reset (void);
 
 	void write_u8 (uint8_t data);
-	void write (uint8_t const*buffer, uint8_t size);
+	void write (uint8_t const*buffer, uint16_t size);
 	void write_hton_u16 (uint16_t data);
 	void write_hton_u32 (uint32_t data);
 
 private:
-	void ensure_write_room_left (uint8_t needed);
+	void ensure_write_room_left (uint16_t needed);
 	uint32_t m_size;
 	uint32_t m_write;
 	uint8_t *m_buffer;
@@ -56,12 +56,12 @@ public:
 	uint8_t *peek_data (void);
 
 	uint8_t read_u8 (void);
-	void read (uint8_t *buffer, uint8_t size);
+	void read (uint8_t *buffer, uint16_t size);
 	uint16_t read_ntoh_u16 (void);
 	uint32_t read_ntoh_u32 (void);
 
 private:
-	bool is_read_room_left (uint8_t needed);
+	bool is_read_room_left (uint16_t needed);
 	uint32_t m_size;
 	uint32_t m_read;
 	uint8_t *m_buffer;
